requirements.txt
python
python-dotenv
langchain
langchain-community
langchain-google-genai
chromadb
langchainhub 
wikipedia
tavily-python
fastapi
uvicorn
sqlalchemy